src/requirements.txt
python-telegram-bot==20.8
boto3==1.34.41